+++ spec/support/gemnasium_helper.rb
@@ -10,10 +10,14 @@ def stub_config(options = {})
   stubbed_config.stub(:site).and_return('gemnasium.com')
   stubbed_config.stub(:use_ssl).and_return(true)
   stubbed_config.stub(:api_key).and_return('test_api_key')
-  stubbed_config.stub(:api_version).and_return('v2')
-  stubbed_config.stub(:profile_name).and_return('tech-angels')
-  stubbed_config.stub(:project_name).and_return(options[:project_name] || 'gemnasium-gem')
+  stubbed_config.stub(:api_version).and_return('v3')
+  stubbed_config.stub(:project_name).and_return(options.fetch(:project_name, 'gemnasium-gem'))
+  stubbed_config.stub(:project_slug).and_return(options.fetch(:project_slug, 'existing-slug'))
   stubbed_config.stub(:project_branch).and_return('master')
+  stubbed_config.stub(:writable?).and_return(options.fetch(:writable?, true))
+  stubbed_config.stub(:store_value!).and_return(true)
+  stubbed_config.stub(:needs_to_migrate?).and_return(options.fetch(:needs_to_migrate?, false))
+  stubbed_config.stub(:migrate!).and_return(nil)
 
   Gemnasium.stub(:config).and_return(stubbed_config)
   Gemnasium.stub(:load_config).and_return(stubbed_config)
@@ -21,52 +25,62 @@ end
 
 def stub_requests
   config = Gemnasium.config
+  request_headers = {'Accept'=>'application/json', 'Content-Type'=>'application/json'}
+  response_headers = {'Content-Type'=>'application/json'}
+
   # Push requests
-  stub_request(:post, api_url("/api/#{config.api_version}/profiles/#{config.profile_name}/projects/up_to_date_project/dependency_files/compare"))
-           .with(:headers => {'Accept'=>'application/json', 'Content-Type'=>'application/json'})
+  stub_request(:post, api_url("/api/#{config.api_version}/projects/up_to_date_project/dependency_files/compare"))
+           .with(:headers => request_headers)
            .to_return(:status => 200,
                       :body => '{ "to_upload": [], "deleted": [] }',
-                      :headers => {'Content-Type'=>'application/json'})
-  stub_request(:post, api_url("/api/#{config.api_version}/profiles/#{config.profile_name}/projects/gemnasium-gem/dependency_files/compare"))
+                      :headers => response_headers)
+
+  stub_request(:post, api_url("/api/#{config.api_version}/projects/existing-slug/dependency_files/compare"))
            .with(:body => '{"new_gemspec.gemspec":"gemspec_sha1_hash","modified_lockfile.lock":"lockfile_sha1_hash","Gemfile_unchanged.lock":"gemfile_sha1_hash"}',
-                 :headers => {'Accept'=>'application/json', 'Content-Type'=>'application/json'})
+                 :headers => request_headers)
            .to_return(:status => 200,
                       :body => '{ "to_upload": ["new_gemspec.gemspec", "modified_lockfile.lock"], "deleted": ["old_dependency_file"] }',
-                      :headers => {'Content-Type'=>'application/json'})
-  stub_request(:post, api_url("/api/#{config.api_version}/profiles/#{config.profile_name}/projects/gemnasium-gem/dependency_files/upload"))
+                      :headers => response_headers)
+
+  stub_request(:post, api_url("/api/#{config.api_version}/projects/existing-slug/dependency_files/upload"))
            .with(:body => '[{"filename":"new_gemspec.gemspec","sha":"gemspec_sha1_hash","content":"stubbed gemspec content"},{"filename":"modified_lockfile.lock","sha":"lockfile_sha1_hash","content":"stubbed lockfile content"}]',
-                 :headers => {'Accept'=>'application/json', 'Content-Type'=>'application/json'})
+                 :headers => request_headers)
            .to_return(:status => 200,
                       :body => '{ "added": ["new_gemspec.gemspec"], "updated": ["modified_lockfile.lockfile"], "unchanged": [], "unsupported": [] }',
-                      :headers => {})
+                      :headers => response_headers)
+
   # Create requests
-  stub_request(:post, api_url("/api/#{config.api_version}/profiles/#{config.profile_name}/projects"))
+  stub_request(:post, api_url("/api/#{config.api_version}/projects"))
           .with(:body => '{"name":"gemnasium-gem","branch":"master"}',
-                :headers => {'Accept'=>'application/json', 'Content-Type'=>'application/json'})
+                :headers => request_headers)
           .to_return(:status => 200,
-                     :body => "{ \"name\": \"gemnasium-gemn\", \"profile\": \"#{config.profile_name}\", \"remaining_slot\": 9001 }",
-                     :headers => {'Content-Type'=>'application/json'})
-  stub_request(:post, api_url("/api/#{config.api_version}/profiles/#{config.profile_name}/projects"))
-          .with(:body => '{"name":"existing_project","branch":"master"}',
-                :headers => {'Accept'=>'application/json', 'Content-Type'=>'application/json'})
-          .to_return do |request|
-            request_body = JSON.parse(request.body)
-            {
-              :status   => 422,
-              :body     => { error: 'project_already_exists', message: "The project `#{request_body['name']}` already exists for the profile `#{config.profile_name}`." }.to_json,
-              :headers  => { 'Content-Type'=>'application/json' }
-            }
-          end
-  stub_request(:post, api_url("/api/#{config.api_version}/profiles/#{config.profile_name}/projects"))
-          .with(:body => '{"name":"existing_project","branch":"master","overwrite_attributes":true}',
-                :headers => {'Accept'=>'application/json', 'Content-Type'=>'application/json'})
-          .to_return(:status => 200,
-                     :body => "{ \"name\": \"existing_project\", \"profile\": \"#{config.profile_name}\", \"remaining_slot\": 9001 }",
-                     :headers => {'Content-Type'=>'application/json'})
+                     # FIXME: make sure the response body is consistent with API v3
+                     :body => '{ "name": "gemnasium-gem", "slug": "new-slug", "remaining_slot_count": 9001 }',
+                     :headers => response_headers)
+
+  # Index requests
+  #
+  # search: offline project, master branch
+  # no-candidate: github project, no match
+  # one-candidate: one project for master branch, another for dev branch, one match
+  # many-candidates: 2 projects matching master branch, 2 matches
+  #
+  stub_request(:get, api_url("/api/#{config.api_version}/projects"))
+          .with(:headers => request_headers)
+          .to_return(:status => 200, :headers => response_headers,
+                     :body => '[
+{"slug":"no-candidate-slug","name":"no-candidate","origin":"github","branch":"master","private":false},
+{"slug":"one-candidate-slug","name":"one-candidate","origin":"offline","branch":"master","private":true},
+{"slug":"one-candidate-slug-dev","name":"one-candidate","origin":"offline","branch":"dev","private":true},
+{"slug":"many-candidates-slug-1","name":"many-candidates","origin":"offline","branch":"master","private":false},
+{"slug":"many-candidates-slug-2","name":"many-candidates","origin":"offline","branch":"master","private":false}
+]'
+                    )
 
   # Connection model's test requests
   stub_request(:get, api_url('/test_path'))
           .with(:headers => {'Accept'=>'application/json', 'Content-Type'=>'application/json'})
+
   stub_request(:post, api_url('/test_path'))
-          .with(:body => {"foo"=>"bar"}, :headers => {'Accept'=>'application/json', 'Content-Type'=>'application/json'})
+          .with(:body => {"foo"=>"bar"}, :headers => request_headers)
 end