+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 3.0.0 / 2013-02-10
+
+**API V2 is now deprecated and all previous gem releases have been yanked**
+
+* Update config file syntax: replace `profile_name` with `project_slug`
+* Add `migrate` command to migrate the config file from previous versions
+* Add `resolve` command to find a project that matches a name and a branch
+* Make it compatible with Gemnasium API v3
+
 # 2.0.2 / 2013-07-31
 
 * [#6][] Fix spec for fedora packaging (@ktdreyer)
@@ -10,8 +19,8 @@
 
 **API V1 is now deprecated and all previous gem releases have been yanked**
 
-* Drop project_visibility option, all offline projects are private now
-* Add ignored_paths options to avoid pushing useless files
+* Drop `project_visibility` option, all offline projects are private now
+* Add `ignored_paths` options to avoid pushing useless files
 
 # 1.0.1 / 2013-04-02